backend/requirements.txt
Django ~= 5.0
psycopg2
django-cors-headers ~= 4.3
djangorestframework ~= 3.14
django-filter ~= 23.5
django-rest-knox==4.*
drf-nested-routers
Pillow ~= 10.2